\name{gnm-defunct}
\alias{gnm-defunct}
\alias{getModelFrame}
\alias{Nonlin}
\title{Defunct Functions in gnm Package}
\description{
  The functions listed here are no longer part of gnm as they are not
  needed any more. 
}
\usage{getModelFrame()}
\usage{Nonlin()}
\details{
  \code{getModelFrame()} and \code{Nonlin} are not needed any more as
  the plug-in architecture has been replaced by functions of class
  \code{"nonlin"}, see \code{\link{nonlin.function}}.
}
\seealso{\code{\link{.Defunct}}}
\keyword{documentation}
