\name{gnm-defunct}
\alias{gnm-defunct}
\alias{Nonlin}
\title{Defunct Functions in gnm Package}
\description{
  The functions listed here are no longer part of gnm as they are not
  needed any more. 
}
\usage{Nonlin()}
\details{
  \code{Nonlin} is not needed any more as the plug-in architecture has
  been replaced by functions of class \code{"nonlin"}, see
  \code{\link{nonlin.function}}.
}
\seealso{\code{\link{.Defunct}}}
\keyword{documentation}
